Refactor and update various components for improved performance and c… (#69)
* Refactor and update various components for improved performance and consistency

- Adjusted emission standard deviations in HMM regime detector for clarity.
- Changed lastVolume type from int64 to uint64 in benchmarkState for better volume handling.
- Cleaned up formatting in GARCH state struct for consistency.
- Enhanced comments in RiskBudgetManager for better understanding of dynamic risk budgeting.
- Standardized volume types from int64 to uint64 in scanner and signals for consistency.
- Removed unused DollarBar and OBV signal implementations to streamline codebase.
- Updated VPIN signal to use uint64 for volume calculations, ensuring consistency across signals.
- Modified main.go to handle Alpaca account values with InexactFloat64 for precision.
- Added MaxSpreadPct configuration to default profile for execution settings.

* refactor: clean up Alpaca client methods and reintroduce data feed logic

* refactor: update equity asset handling to use Alpaca's Asset type

+++ historical_cache_codec.go
@@ -141,7 +141,7 @@ func (r *historicalJobCacheReader) Next() (backtest.InputBar, bool, error) {
 		High:      decodeHistoricalPrice(highPrice),
 		Low:       decodeHistoricalPrice(lowPrice),
 		Close:     decodeHistoricalPrice(closePrice),
-		Volume:    int64(volume),
+		Volume:    volume,
 		PrevClose: decodeHistoricalPrice(prevClose),
 	}, true, nil
 }
